PolicyVault-Nexus-Real-Time-RBI-Compliant-Data-Gateway-main/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.3.4
cryptography==45.0.5
python-dotenv==1.1.1
pydantic==2.11.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
redis==5.2.1
httpx==0.28.1
pyotp==2.9.0
starlette==0.47.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
scipy==1.15.3
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.2.6
